main/serializers.py
from rest_framework import serializers
from .models import CrimeReport, User

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['id', 'username', 'role', 'email']

class CrimeReportSerializer(serializers.ModelSerializer):
    reporter = UserSerializer(read_only=True)

    class Meta:
        model = CrimeReport
        fields = '__all__'
