requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
distro==1.9.0
Django==5.2.8
djangorestframework==3.16.1
geographiclib==2.1
geopy==2.4.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
openai==2.7.1
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
pydantic==2.12.4
pydantic_core==2.41.5
python-decouple==3.8
python-dotenv==1.2.1
reportlab==4.4.4
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
